Enquiry notification added

+++ public/js/app.b250529ef438ebbe526c.js
@@ -4780,7 +4780,7 @@ Object.defineProperty(__webpack_exports__, "__esModule", { value: true });
         return {
             isLoading: false,
             page: 0,
-            limit: 16,
+            limit: 36,
             portfolios: [],
             finished: false
         };
@@ -96713,18 +96713,23 @@ if ($('.blog-excerpt').length > 0) {
 			var id = $(this).data('id');
 			var data = { _method: 'delete' };
 			$('body').addClass('is-loading');
-			axios.delete('/dashboard/blog/' + id + '/delete', data).then(function (response) {
-				if (response.data === true) {
-					$(el).closest('.blog-excerpt').remove();
-					iziToast.success({ title: 'Blog post deleted successfully.' });
-				} else {
-					iziToast.error({ title: 'Error occured while deleting blog post.' });
-				}
-				$('body').removeClass('is-loading');
-				// console.log(response);
-			}).catch(function (error) {
-				return $('body').removeClass('is-loading');
-			});
+			var confirm = window.confirm("Are you sure you want to delete this blog post? This action cannot be undone");
+			if (confirm) {
+				axios.delete('/dashboard/blog/' + id + '/delete', data).then(function (response) {
+					if (response.data === true) {
+						$(el).closest('.blog-excerpt').remove();
+						iziToast.success({ title: 'Blog post deleted successfully.' });
+					} else {
+						iziToast.error({ title: 'Error occured while deleting blog post.' });
+					}
+					$('body').removeClass('is-loading');
+					// console.log(response);
+				}).catch(function (error) {
+					return $('body').removeClass('is-loading');
+				});
+			} else {
+				return;
+			}
 		});
 	});
 }
